requirements.txt
PyQt5>=5.15
pocketflow>=0.0.1
google-generativeai>=0.5.0
openai>=1.3.0
zai-sdk
pytubefix>=5.0.0
youtube-transcript-api>=0.6.0
yt-dlp>=2023.12.30
ffmpeg-python>=0.2.0
python-dotenv>=1.0.0
requests>=2.25.0
feedparser>=6.0.0
audioop-lts>=0.2.1; python_version >= "3.13"
curl_cffi>=0.5.0
markitdown[all]>=0.1.0
extract-msg>=0.46.0   # .msg email body extraction (MarkItDown often returns only header)

# Notes:
# - pathlib is part of the standard library for Python 3.4+
# - audioop-lts: Required for Python 3.13+ (audioop module was removed in PEP 594)
# 
# Not needed (commented out):
# - openai-whisper: Project uses OpenAI API for transcription, not local whisper model
# - pydub: Project uses ffmpeg-python directly for audio processing